+++ catwalk/steps.py
@@ -9,6 +9,7 @@ from typing import (
     Tuple,
     MutableSequence,
 )
+from collections import defaultdict
 
 import tango.integrations.torch
 from tango import Step, JsonFormat
@@ -26,6 +27,7 @@ from tango.integrations.torch import (
     DataLoader,
 )
 from torch.optim import AdamW
+import torch
 
 from catwalk.task import Task
 from catwalk.tasks import TASKS
@@ -163,11 +165,22 @@ class TabulateMetricsStep(Step):
     FORMAT = TextFormat
 
     def run(self, metrics: Dict[str, Dict[str, float]], format: str = "text") -> Iterable[str]:
+        flattend_metrics: Dict[str, Dict[str, float]] = defaultdict(dict)
+        for task_name, task_metrics in metrics.items():
+            for metric_name, metric_value in task_metrics.items():
+                # if metric_value is a dict, then it's a nested metric
+                if isinstance(metric_value, dict):
+                    for nested_metric_name, nested_metric_value in metric_value.items():
+                        flattend_metrics[task_name][f"{metric_name}.{nested_metric_name}"] = nested_metric_value.item() if isinstance(nested_metric_value, torch.Tensor) else nested_metric_value
+                else:
+                    flattend_metrics[task_name][metric_name] = metric_value
+            
         if format == "text":
-            for task_name, task_metrics in metrics.items():
+            for task_name, task_metrics in flattend_metrics.items():
                 for metric_name, metric_value in task_metrics.items():
                     yield f"{task_name}\t{metric_name}\t{metric_value}"
         elif format == "latex":
             raise NotImplementedError()
         else:
             raise AttributeError("At the moment, only the 'text' format is supported.")
+